LFS-613: Fix questionnaire deletion and remove duplicate Actions column - Fix recursive deletion of questionnaires   - Nodes with mandatory child nodes could not be deleted recursively - Remove second actions column from admin Questionnaire page   - Remove questionnaire specific ddelete button in favor of generic delete button   - Merge log in behaviour from questionnaire deletion button to generic button   - Fix error handling in generic delete button to inform user of errors
