+++ modules/data-entry/src/main/frontend/src/dataHomepage/Questionnaires.jsx
@@ -21,27 +21,14 @@ import LiveTable from "./LiveTable.jsx";
 import Questionnaire from "../questionnaire/Questionnaire.jsx";
 import QuestionnaireStyle from "../questionnaire/QuestionnaireStyle";
 import NewQuestionnaireDialog from "../questionnaireEditor/NewQuestionnaireDialog.jsx";
-import DeleteQuestionnaireDialog from "../questionnaireEditor/DeleteQuestionnaireDialog.jsx";
-import { Button, Card, CardHeader, CardContent, Typography, withStyles } from "@material-ui/core";
+import { Button, Card, CardHeader, CardContent, IconButton, Tooltip, Typography, withStyles } from "@material-ui/core";
+import { Link } from 'react-router-dom';
+import EditIcon from "@material-ui/icons/Edit";
 import DeleteButton from "./DeleteButton.jsx";
 
 function Questionnaires(props) {
-  let [ openDialog, setOpenDialog ] = useState(false);
-  let [ dataToDelete, setDataToDelete ] = useState(false);
-  let [ error, setError ] = useState();
-  let [ deletionCount, setDeletionCount ] = useState(0); // Used to force reupdates to the livetable after deletion
   const { classes } = props;
   const entry = /Questionnaires\/(.+)/.exec(location.pathname);
-
-  let deleteQuestionnaire = (questionnaire) => {
-    setDataToDelete(questionnaire);
-    setOpenDialog(true);
-  }
-
-  // Increment the numebr of times we have deleted something, forcing the LiveTable to update
-  let updateDeletionCount = () => {
-    setDeletionCount((old) => (old+1));
-  }
 
   if (entry) {
     return <Questionnaire id={entry[1]} key={location.pathname}/>;
@@ -64,21 +51,22 @@ function Questionnaires(props) {
       "key": "jcr:created",
       "label": "Created on",
       "format": "date:YYYY-MM-DD HH:mm",
-    },
-    {
-      "key":"actions",
-      "label":"Actions",
-      "admin": true,
-      // Align the actions to the end
-      "props": {
-        "style": {
-          "textAlign": "end"
-        }
-      }
     }
   ]
   const actions = [
-    DeleteButton
+    DeleteButton,
+    function EditButton(props) {
+      const { entryPath, entryType, buttonClass } = props;
+      return(
+        <Link to={"/content.html/admin" + entryPath}>
+          <Tooltip title={entryType ? "Edit " + entryType : "Edit"}>
+            <IconButton className={buttonClass}>
+              <EditIcon />
+            </IconButton>
+          </Tooltip>
+        </Link>
+      )
+    }
   ]
 
   return (
@@ -101,25 +89,11 @@ function Questionnaires(props) {
         <CardContent>
           <LiveTable
             columns={columns}
-            delete={deleteQuestionnaire}
-            updateData={deletionCount}
             actions={actions}
             entryType={"Questionnaire"}
             />
-          { error &&
-            <Typography color="error" variant="h3">
-              {error}
-            </Typography>
-          }
         </CardContent>
       </Card>
-      <DeleteQuestionnaireDialog
-        open={openDialog}
-        onClose={() => {setOpenDialog(false);}}
-        onDelete={updateDeletionCount}
-        data={dataToDelete}
-        onError={setError}
-        />
     </React.Fragment>
   );
 }
